removed unnecessary code duplication between MavenProcessFactory and Maven3ProcessFactory
