+++ maven-plugin/src/main/java/hudson/maven/Maven3ProcessFactory.java
@@ -27,14 +27,11 @@ import hudson.EnvVars;
 import hudson.FilePath;
 import hudson.Launcher;
 import hudson.model.BuildListener;
-import jenkins.model.Jenkins;
-import hudson.model.JDK;
 import hudson.model.Run.RunnerAbortedException;
 import hudson.model.TaskListener;
 import hudson.remoting.Callable;
 import hudson.remoting.Which;
 import hudson.tasks.Maven.MavenInstallation;
-import hudson.util.ArgumentListBuilder;
 
 import java.io.File;
 import java.io.FilenameFilter;
@@ -53,75 +50,32 @@ public class Maven3ProcessFactory extends AbstractMavenProcessFactory implements
     Maven3ProcessFactory(MavenModuleSet mms, Launcher launcher, EnvVars envVars, String mavenOpts, FilePath workDir) {
         super( mms, launcher, envVars, mavenOpts, workDir );
     }
-    /**
-     * Builds the command line argument list to launch the maven process.
-     *
-     */
-    protected ArgumentListBuilder buildMavenAgentCmdLine(BuildListener listener,int tcpPort) throws IOException, InterruptedException {
-        MavenInstallation mvn = getMavenInstallation(listener);
-        if(mvn==null) {
-            listener.error("Maven version is not configured for this project. Can't determine which Maven to run");
-            throw new RunnerAbortedException();
-        }
-        if(mvn.getHome()==null) {
-            listener.error("Maven '%s' doesn't have its home set",mvn.getName());
-            throw new RunnerAbortedException();
-        }
 
-        // find classworlds.jar
+    @Override
+    protected String getMavenAgentClassPath(MavenInstallation mvn,boolean isMaster,FilePath slaveRoot,BuildListener listener) throws IOException, InterruptedException {
         String classWorldsJar = getLauncher().getChannel().call(new GetClassWorldsJar(mvn.getHome(),listener));
-
-        boolean isMaster = getCurrentNode()== Jenkins.getInstance();
-        FilePath slaveRoot=null;
-        if(!isMaster)
-            slaveRoot = getCurrentNode().getRootPath();
-
-        ArgumentListBuilder args = new ArgumentListBuilder();
-        JDK jdk = getJava(listener);
-        if(jdk==null) {
-            args.add("java");
-        } else {
-            args.add(jdk.getHome()+"/bin/java"); // use JDK.getExecutable() here ?
-        }
-
-        if(MavenProcessFactory.debugPort!=0)
-            args.add("-Xrunjdwp:transport=dt_socket,server=y,address="+MavenProcessFactory.debugPort);
-        if(MavenProcessFactory.yjp)
-            args.add("-agentlib:yjpagent=tracing");
-
-        args.addTokenized(getMavenOpts());
         
-        args.add("-cp");
-        args.add(
-            (isMaster? Which.jarFile(Maven3Main.class).getAbsolutePath():slaveRoot.child("maven3-agent.jar").getRemote())+
-            (getLauncher().isUnix()?":":";")+classWorldsJar);
-        args.add(Maven3Main.class.getName());
-
-        // M2_HOME
-        args.add(mvn.getHome());
-
-        // remoting.jar
-        String remotingJar = getLauncher().getChannel().call(new GetRemotingJar());
-        if(remotingJar==null) {// this shouldn't be possible, but there are still reports indicating this, so adding a probe here.
-            listener.error("Failed to determine the location of slave.jar");
-            throw new RunnerAbortedException();
-        }
-        args.add(remotingJar);
-        
-        args.add(isMaster?
-            Which.jarFile(Maven3Launcher.class).getAbsolutePath():
-            slaveRoot.child("maven3-interceptor.jar").getRemote());
-
-        // TCP/IP port to establish the remoting infrastructure
-        args.add(tcpPort);
-        
-        return args;
+        return (isMaster? Which.jarFile(Maven3Main.class).getAbsolutePath():slaveRoot.child("maven3-agent.jar").getRemote())+
+            (getLauncher().isUnix()?":":";")+classWorldsJar;
+    }
+    
+    @Override
+    protected String getMainClassName() {
+        return Maven3Main.class.getName();
+    }
+    
+    @Override
+    protected String getMavenInterceptorClassPath(MavenInstallation mvn,boolean isMaster,FilePath slaveRoot) throws IOException, InterruptedException {
+        return isMaster?
+                Which.jarFile(Maven3Launcher.class).getAbsolutePath():
+                slaveRoot.child("maven3-interceptor.jar").getRemote();
     }
 
     /**
      * Finds classworlds.jar
      */
     private static final class GetClassWorldsJar implements Callable<String,IOException> {
+        private static final long serialVersionUID = -2599434124883557137L;
         private final String mvnHome;
         private final TaskListener listener;
 
@@ -137,15 +91,8 @@ public class Maven3ProcessFactory extends AbstractMavenProcessFactory implements
             File bootDir = new File(home, "boot");
             File[] classworlds = bootDir.listFiles(CLASSWORLDS_FILTER);
             if(classworlds==null || classworlds.length==0) {
-                // Maven 2.0.6 puts it to a different place
-                bootDir = new File(home, "boot");
-                classworlds = bootDir.listFiles(CLASSWORLDS_FILTER);
-                if(classworlds==null || classworlds.length==0) {
-                    // FIXME use messages
-                    //listener.error(Messages.MavenProcessFactory_ClassWorldsNotFound(home));
-                    listener.error("classworld not found");
-                    throw new RunnerAbortedException();
-                }
+                listener.error(Messages.MavenProcessFactory_ClassWorldsNotFound(home));
+                throw new RunnerAbortedException();
             }
             return classworlds[0].getAbsolutePath();
         }
