+++ maven-plugin/src/main/java/hudson/maven/MavenProcessFactory.java
@@ -29,14 +29,11 @@ import hudson.EnvVars;
 import hudson.maven.agent.Main;
 import hudson.maven.agent.Maven21Interceptor;
 import hudson.model.BuildListener;
-import jenkins.model.Jenkins;
-import hudson.model.JDK;
 import hudson.model.Run.RunnerAbortedException;
 import hudson.model.TaskListener;
 import hudson.remoting.Callable;
 import hudson.remoting.Which;
 import hudson.tasks.Maven.MavenInstallation;
-import hudson.util.ArgumentListBuilder;
 
 import java.io.File;
 import java.io.FilenameFilter;
@@ -55,88 +52,39 @@ final class MavenProcessFactory extends AbstractMavenProcessFactory implements P
         super( mms, launcher, envVars, mavenOpts, workDir );
     }
 
-    /**
-     * Builds the command line argument list to launch the maven process.
-     *
-     * UGLY.
-     */
-    protected ArgumentListBuilder buildMavenAgentCmdLine(BuildListener listener,int tcpPort) throws IOException, InterruptedException {
-        MavenInstallation mvn = getMavenInstallation(listener);
-        if(mvn==null) {
-            listener.error("Maven version is not configured for this project. Can't determine which Maven to run");
-            throw new RunnerAbortedException();
-        }
-        if(mvn.getHome()==null) {
-            listener.error("Maven '%s' doesn't have its home set",mvn.getName());
-            throw new RunnerAbortedException();
-        }
-
-        // find classworlds.jar
+    @Override
+    protected String getMavenAgentClassPath(MavenInstallation mvn,boolean isMaster,FilePath slaveRoot,BuildListener listener) throws IOException, InterruptedException {
         String classWorldsJar = getLauncher().getChannel().call(new GetClassWorldsJar(mvn.getHome(),listener));
-
-        boolean isMaster = getCurrentNode()== Jenkins.getInstance();
-        FilePath slaveRoot=null;
-        if(!isMaster)
-            slaveRoot = getCurrentNode().getRootPath();
-
-        ArgumentListBuilder args = new ArgumentListBuilder();
-        JDK jdk = getJava(listener);
-        if(jdk==null) {
-            args.add("java");
-        } else {
-            args.add(jdk.getHome()+"/bin/java"); // use JDK.getExecutable() here ?
-        }
-
-        if(debugPort!=0)
-            args.add("-Xrunjdwp:transport=dt_socket,server=y,address="+debugPort);
-        if(yjp)
-            args.add("-agentlib:yjpagent=tracing");
-
-        args.addTokenized(getMavenOpts());
-        
-        args.add( "-cp" );
         String classPath =
             ( isMaster ? Which.jarFile( Main.class ).getAbsolutePath()
                             : slaveRoot.child( "maven-agent.jar" ).getRemote() )
                 + ( getLauncher().isUnix() ? ":" : ";" )
                 + ( isMaster ? classWorldsJar : slaveRoot.child( "classworlds.jar" ).getRemote() );
-        args.add( classPath );
-            //+classWorldsJar);
-        args.add(Main.class.getName());
-
-        // M2_HOME
-        args.add(mvn.getHome());
-
-        // remoting.jar
-        String remotingJar = getLauncher().getChannel().call(new GetRemotingJar());
-        if(remotingJar==null) {// this shouldn't be possible, but there are still reports indicating this, so adding a probe here.
-            listener.error("Failed to determine the location of slave.jar");
-            throw new RunnerAbortedException();
-        }
-        args.add(remotingJar);
-
-        // interceptor.jar
-        args.add(isMaster?
-            Which.jarFile(hudson.maven.agent.AbortException.class).getAbsolutePath():
-            slaveRoot.child("maven-interceptor.jar").getRemote());
-
-        // TCP/IP port to establish the remoting infrastructure
-        args.add(tcpPort);
-
-        // if this is Maven 2.1, interceptor override
-        if(mvn.isMaven2_1(getLauncher())) {
-            args.add(isMaster?
-                Which.jarFile(Maven21Interceptor.class).getAbsolutePath():
-                slaveRoot.child("maven2.1-interceptor.jar").getRemote());
-        }
-       
-        return args;
+        return classPath;
     }
     
+    protected String getMainClassName() {
+        return Main.class.getName();
+    }
+
+    @Override
+    protected String getMavenInterceptorClassPath(MavenInstallation mvn,boolean isMaster,FilePath slaveRoot) throws IOException, InterruptedException {
+        if(mvn.isMaven2_1(getLauncher())) {
+            return isMaster?
+                Which.jarFile(Maven21Interceptor.class).getAbsolutePath():
+                slaveRoot.child("maven2.1-interceptor.jar").getRemote();
+        } else {
+            return isMaster?
+                Which.jarFile(hudson.maven.agent.AbortException.class).getAbsolutePath():
+                slaveRoot.child("maven-interceptor.jar").getRemote();
+        }
+    }
+
     /**
      * Finds classworlds.jar
      */
     private static final class GetClassWorldsJar implements Callable<String,IOException> {
+        private static final long serialVersionUID = 5812919424079344101L;
         private final String mvnHome;
         private final TaskListener listener;
 
@@ -193,23 +141,7 @@ final class MavenProcessFactory extends AbstractMavenProcessFactory implements P
      */
     public static boolean debug = false;
 
-    /**
-     * If not 0, launch Maven with a debugger port.
-     */
-    public static int debugPort;
-
     public static boolean profile = Boolean.getBoolean("hudson.maven.profile");
-    
-    /**
-     * If true, launch Maven with YJP offline profiler agent.
-     */
-    public static boolean yjp = Boolean.getBoolean("hudson.maven.yjp");
-
-    static {
-        String port = System.getProperty("hudson.maven.debugPort");
-        if(port!=null)
-            debugPort = Integer.parseInt(port);
-    }
     
     public static int socketTimeOut = Integer.parseInt( System.getProperty( "hudson.maven.socketTimeOut", Integer.toString( 30*1000 ) ) );
 }
